requirements.txt
asttokens==2.4.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.10.0
wcwidth==0.2.13
openai==1.10.0
notebook
python-dotenv
ipykernel